Unification of ADCS task; improved linalg logic; updated command structure

+++ src/linalg/LinearAlgebra/diagpower.c
@@ -1,5 +1,5 @@
 /*
- * sqrte.c
+ * sqrtfe.c
  *
  *  Created on: 12 feb. 2019
  *      Author: Daniel Mårtensson
@@ -11,15 +11,14 @@
  * Take power p, on all element on the diagonal of matrix A, who has the size row x column
  */
 
-void diagpower(double* A, double p, int row, int column){
+void diagpower(float* A, float p, int row, int column) {
 
-	for(int i = 0; i < row; i++){
-		for(int j = 0; j < column; j++){
-			if(j == i){
-				*(A) = powf(*(A), p);
-				A += column + 1;
-			}
-		}
-	}
+    for (int i = 0; i < row; i++) {
+        for (int j = 0; j < column; j++) {
+            if (j == i) {
+                *(A) = powf(*(A), p);
+                A += column + 1;
+            }
+        }
+    }
 }
-
