+++ src/linalg/LinearAlgebra/move.c
@@ -12,25 +12,23 @@
  * Matrix A have the size row x column
  */
 
-void move(double* A, int row, int column, int down, int right) {
+void move(float* A, int row, int column, int down, int right) {
 
-	double B[row*column];
-	memset(B, 0, row*column*sizeof(double));
+    float B[row * column];
+    memset(B, 0, row * column * sizeof(float));
 
-	/*
-	 * We take -1 due to indexing from zero
-	 */
+    /*
+     * We take -1 due to indexing from zero
+     */
 
-	for (int i = row-1; i >= 0; i--){
-		for (int j = column-1; j >= 0; j--){
-			if(i -down <= -1)
-				*(B + i*column + j) = 0;
-			else
-				if(j - right >= 0)
-					*(B + i*column + j) = *(A + (i-down)*column + j - right);
-		}
-	}
+    for (int i = row - 1; i >= 0; i--) {
+        for (int j = column - 1; j >= 0; j--) {
+            if (i - down <= -1)
+                *(B + i * column + j) = 0;
+            else if (j - right >= 0)
+                *(B + i * column + j) = *(A + (i - down) * column + j - right);
+        }
+    }
 
-
-	memcpy(A, B, row*column*sizeof(double));
+    memcpy(A, B, row * column * sizeof(float));
 }
