+++ src/linalg/LinearAlgebra/sumrows.c
@@ -11,12 +11,11 @@
  * Sum all rows from A size row x column, into a flat matrix B size 1 x column
  */
 
-void sumrows(double* A, double* B, int row, int column) {
+void sumrows(float* A, float* B, int row, int column) {
 
-	for(int j = 0; j < column; j++){
-		for(int i = 0; i < row; i++){
-			*(B + j) += *((A + i*column) + j);
-		}
-	}
-
+    for (int j = 0; j < column; j++) {
+        for (int i = 0; i < row; i++) {
+            *(B + j) += *((A + i * column) + j);
+        }
+    }
 }
