+++ src/linalg/LinearAlgebra/vertcat.c
@@ -14,24 +14,23 @@
  * C size (row_a + row_b) x column_a
  */
 
-void vertcat(double* A, double* B, double* C, int row_a, int column_a, int row_b) {
+void vertcat(float* A, float* B, float* C, int row_a, int column_a, int row_b) {
 
-	/*
-	 * C = [A; B]
-	 */
+    /*
+     * C = [A; B]
+     */
 
-	// We include the first matrix a
-	for (int i = 0; i < row_a; i++) {
-		for (int j = 0; j < column_a; j++) {
-			*(C++) = *((A + i * column_a) + j);
-		}
-	}
+    // We include the first matrix a
+    for (int i = 0; i < row_a; i++) {
+        for (int j = 0; j < column_a; j++) {
+            *(C++) = *((A + i * column_a) + j);
+        }
+    }
 
-
-	// We change the matrix now - We use b matrix now
-	for (int i = 0; i < row_b; i++) {
-		for (int j = 0; j < column_a; j++) {
-			*(C++) = *((B + i * column_a) + j);
-		}
-	}
+    // We change the matrix now - We use b matrix now
+    for (int i = 0; i < row_b; i++) {
+        for (int j = 0; j < column_a; j++) {
+            *(C++) = *((B + i * column_a) + j);
+        }
+    }
 }
